requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.txt
#
requests==2.22.0
django-cors-headers==3.1.0
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
coverage==4.5.4
dj-database-url==0.5.0
django-extensions==2.1.6
django-heroku==0.3.1
django-nose==1.4.6
django-rest-framework==0.1.0
django==2.1.7
djangorestframework==3.9.2
gunicorn==19.9.0
idna==2.8
nose==1.3.7
pinocchio==0.4.2
psycopg2
pyjwt==1.7.1
pytz==2018.9
six==1.12.0
urllib3==1.25.3
whitenoise==4.1.2
